requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
dash==1.9.0
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Events==0.3
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
googleapis-common-protos==1.6.0
GPy==1.9.8
grpcio==1.22.0
idna==2.8
influxdb==5.2.2
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.17.0
paho-mqtt==1.4.0
pandas==0.25.0
pandocfilters==1.4.2
paramz==0.9.5
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plotly==4.5.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.2
requests==2.22.0
retrying==1.3.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
ttn==2.1.5
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
thingspeak==1.0.0